requirements.txt
Babel==1.3
-e  git+https://github.com/mitsuhiko/flask.git#egg=flask
Flask-Babel==0.8
Flask-Login==0.2.11
Flask-Mail==0.7.6
Flask-OpenID==1.2.4
Flask-Restless==0.13.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.8.4
Flask-WhooshAlchemy==0.56
Jinja2==2.7
Mako==1.0.0
MarkupSafe==0.18
SQLAlchemy==0.9.8
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.2
Whoosh==2.6.0
alembic==0.6.7
blinker==1.3
decorator==3.4.0
defusedxml==0.4.1
gunicorn==17.5
itsdangerous==0.22
mimerender==0.5.4
pbr==0.10.0
python-dateutil==2.1
python-mimeparse==0.1.4
python3-openid==3.0.5
pytz==2013b
requests==1.2.3
six==1.3.0
speaklater==1.3
sqlalchemy-migrate==0.9.2
sqlparse==0.1.13
